+++ Kernel/Modules/AgentTicketLockedView.pm
@@ -175,6 +175,8 @@ sub Run {
         $SortByS = 'Age';
     }
 
+    my $UserIDForSearch = $Config->{TicketSearchWithAdminUser} ? 1 : $Self->{UserID};
+
     my %Filters = (
         All => {
             Name   => Translatable('All'),
@@ -184,7 +186,7 @@ sub Run {
                 OwnerIDs   => [ $Self->{UserID} ],
                 OrderBy    => $OrderBy,
                 SortBy     => $SortByS,
-                UserID     => $Self->{UserID},
+                UserID     => $UserIDForSearch,
                 Permission => 'ro',
             },
         },
@@ -200,7 +202,7 @@ sub Run {
                 TicketFlagUserID => $Self->{UserID},
                 OrderBy          => $OrderBy,
                 SortBy           => $SortByS,
-                UserID           => $Self->{UserID},
+                UserID           => $UserIDForSearch,
                 Permission       => 'ro',
             },
         },
@@ -213,7 +215,7 @@ sub Run {
                 OwnerIDs   => [ $Self->{UserID} ],
                 OrderBy    => $OrderBy,
                 SortBy     => $SortByS,
-                UserID     => $Self->{UserID},
+                UserID     => $UserIDForSearch,
                 Permission => 'ro',
             },
         },
@@ -227,7 +229,7 @@ sub Run {
                 OwnerIDs                      => [ $Self->{UserID} ],
                 OrderBy                       => $OrderBy,
                 SortBy                        => $SortByS,
-                UserID                        => $Self->{UserID},
+                UserID                        => $UserIDForSearch,
                 Permission                    => 'ro',
             },
         },