+++ Kernel/Modules/AgentTicketOwnerView.pm
@@ -170,6 +170,8 @@ sub Run {
         $SortByS = 'Age';
     }
 
+    my $UserIDForSearch = $Config->{TicketSearchWithAdminUser} ? 1 : $Self->{UserID};
+
     my %Filters = (
         All => {
             Name   => Translatable('All'),
@@ -179,7 +181,7 @@ sub Run {
                 OwnerIDs   => [ $Self->{UserID} ],
                 OrderBy    => $OrderBy,
                 SortBy     => $SortByS,
-                UserID     => $Self->{UserID},
+                UserID     => $UserIDForSearch,
                 Permission => 'ro',
             },
         },
@@ -195,7 +197,7 @@ sub Run {
                 TicketFlagUserID => $Self->{UserID},
                 OrderBy          => $OrderBy,
                 SortBy           => $SortByS,
-                UserID           => $Self->{UserID},
+                UserID           => $UserIDForSearch,
                 Permission       => 'ro',
             },
         },
@@ -207,7 +209,7 @@ sub Run {
                 OwnerIDs   => [ $Self->{UserID} ],
                 OrderBy    => $OrderBy,
                 SortBy     => $SortByS,
-                UserID     => $Self->{UserID},
+                UserID     => $UserIDForSearch,
                 Permission => 'ro',
             },
         },
@@ -220,7 +222,7 @@ sub Run {
                 OwnerIDs                      => [ $Self->{UserID} ],
                 OrderBy                       => $OrderBy,
                 SortBy                        => $SortByS,
-                UserID                        => $Self->{UserID},
+                UserID                        => $UserIDForSearch,
                 Permission                    => 'ro',
             },
         },