+++ Kernel/Output/HTML/ToolBar/TicketLocked.pm
@@ -39,15 +39,17 @@ sub Run {
 
     return if !$Kernel::OM->Get('Kernel::Config')->Get('Frontend::Module')->{AgentTicketLockedView};
 
-    # get ticket object
     my $TicketObject = $Kernel::OM->Get('Kernel::System::Ticket');
+
+    my $TicketFrontendConfig = $Kernel::OM->Get('Kernel::Config')->Get('Ticket::Frontend::AgentTicketLockedView') // {};
+    my $UserIDForSearch      = $TicketFrontendConfig->{TicketSearchWithAdminUser} ? 1 : $Self->{UserID};
 
     # get user lock data
     my $Count = $TicketObject->TicketSearch(
         Result     => 'COUNT',
         Locks      => [ 'lock', 'tmp_lock' ],
         OwnerIDs   => [ $Self->{UserID} ],
-        UserID     => $Self->{UserID},
+        UserID     => $UserIDForSearch,
         Permission => 'ro',
     ) || 0;
     my $CountNew = $TicketObject->TicketSearch(
@@ -58,7 +60,7 @@ sub Run {
             Seen => 1,
         },
         TicketFlagUserID => $Self->{UserID},
-        UserID           => $Self->{UserID},
+        UserID           => $UserIDForSearch,
         Permission       => 'ro',
     ) || 0;
     $CountNew = $Count - $CountNew;
@@ -68,7 +70,7 @@ sub Run {
         StateType                     => ['pending reminder'],
         TicketPendingTimeOlderMinutes => 1,
         OwnerIDs                      => [ $Self->{UserID} ],
-        UserID                        => $Self->{UserID},
+        UserID                        => $UserIDForSearch,
         Permission                    => 'ro',
     ) || 0;